requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.5
Flask-CORS==4.0.0
pyserial==3.5
joblib==1.3.2
numpy==1.26.2
scikit-learn==1.3.2
python-socketio==5.10.0
simple-websocket==1.0.0
